\name{apisensr-package}
\alias{apisensr-package}
\alias{apisensr}
\docType{package}
\title{Interface to episensr for sensitivity analysis of epidemiological results}
\description{
  API for using 'episensr' which provides basic sensitivity analysis of
  the observed relative risks adjusting for unmeasured confounding and
  misclassification of the exposure/outcome, or both.}
\author{
Maintainer: Denis Haine \email{denis.haine@gmail.com}
}
\seealso{
  Useful links:
  \itemize{
    \item \url{https://github.com/dhaine/apisensr}
    \item Report bugs at \url{https://github.com/dhaine/apisensr/issues}
    \item episensr on R CRAN \url{https://cran.r-project.org/package=episensr}
  }
}
\examples{
\dontrun{
    run_app()
}
}
