% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_draw_dag.R
\name{draw_mdag_before}
\alias{draw_mdag_before}
\alias{draw_mdag_after}
\title{Draw a M-bias DAG}
\usage{
draw_mdag_before(x, dec = 2, ...)

draw_mdag_after(x, dec = 2, ...)
}
\arguments{
\item{x}{'mbias' object to plot.}

\item{dec}{Number of digits displayed.}

\item{...}{Other unused arguments.}
}
\value{
A DAG for selection bias caused by M bias.
}
\description{
Draw a Directed Acyclic Graph (DAG), either before or after conditioning on the
collider M, for selection bias caused by M bias.
}
