% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{mod_about_ui}
\alias{mod_about_ui}
\alias{mod_about_server}
\title{UI Module for the About Analysis tab}
\usage{
mod_about_ui(id, label = "tab_about")

mod_about_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to render the About analysis tab, including the table
for available functions and comparison with `episensr`, as well as `episensr` version.
}
\keyword{internal}
