% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis.R
\name{mod_analysis_ui}
\alias{mod_analysis_ui}
\alias{mod_analysis_server}
\title{UI Module for Analysis tab}
\usage{
mod_analysis_ui(id, label = "tab_analysis")

mod_analysis_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to render the analysis tab.
}
\keyword{internal}
