% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_notable.R
\name{mod_notable_ui}
\alias{mod_notable_ui}
\alias{mod_notable_server}
\title{UI Module for Analysis tab with no 2x2 table.}
\usage{
mod_notable_ui(id, label = "tab_notable")

mod_notable_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to render the analysis tab when no 2-by-2 table
is provided.
}
\keyword{internal}
