% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_parms.R
\name{mod_parms_ui}
\alias{mod_parms_ui}
\alias{mod_parms_server}
\title{Module to provide simple bias parameters}
\usage{
mod_parms_ui(id, label_parms, value)

mod_parms_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to treat simple Se/Sp bias analysis parameters
(selection, misclassification).
}
\keyword{internal}
