% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prob.R
\name{mod_prob_ui}
\alias{mod_prob_ui}
\alias{mod_prob_server}
\title{UI Module for Probabilistic Analysis tab}
\usage{
mod_prob_ui(id, label = "tab_prob")

mod_prob_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to render the Probabilistic Analysis tab (selection bias
analysis `probsens.sel`, bias analysis for unmeasured confounder `probsens.conf`, and
misclassification bias analysis `probsens`).
}
\keyword{internal}
