% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Unused arguments.}
}
\description{
Runs the 'apisensr' Shiny application. This function does not
return; interrupt R to stop the application (usually by pressing Ctrl+C or Esc).
}
\examples{
\dontrun{
run_app()
}

}
