% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisensr-package.R
\docType{package}
\name{apisensr-package}
\alias{apisensr}
\alias{apisensr-package}
\title{Interface to 'episensr' for sensitivity analysis of epidemiological results}
\description{
API for using 'episensr' which provides basic sensitivity analysis
of the observed relative risks adjusting for unmeasured confounding and
misclassification of the exposure/outcome, or both.
}
\examples{
\dontrun{
run_app()
}
}
\references{
'episensr' on R CRAN \url{https://cran.r-project.org/package=episensr}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dhaine/apisensr}
  \item Report bugs at \url{https://github.com/dhaine/apisensr/issues}
}

}
\author{
\strong{Maintainer}: Denis Haine \email{denis.haine@gmail.com} (\href{https://orcid.org/0000-0002-6691-7335}{ORCID})

}
\keyword{internal}
